requirements.txt
flask==3.1.0
flask-cors==5.0.1
pyyaml==6.0.2
python-dotenv==1.1.0
gunicorn==23.0.0
pytest==8.3.4
pytest-cov==6.0.0
